go production

+++ app/controllers/providers_controller.rb
@@ -8,8 +8,13 @@ class ProvidersController < ApplicationController
   end
   
   def destroy
+    person = Person.find params[:person_id]
+    if ( !person.email || person.email == '') && person.providers.count == 1
+      redirect_to(edit_person_path(person.id), :alert => I18n.t('alert.your_action_might_lead_you_cant_login_anymore'))
+      return
+    end
     provider = Provider.find params[:id]
     provider.destroy
-    redirect_to edit_person_path(params[:person_id])
+    redirect_to edit_person_path(person.id)
   end
 end
